Making name of method to add on User model consistent
Previously in the README the method to add on the User model was noted as `User::getUserEntityByRequest()` in the setup/how to use section, however the example and package source has the method as `User::byPassportCustomRequest()`. This pull request updates the method names to be consistent with the package source code.